edu_video_pipeline/requirements.txt
# Core dependencies
PyMuPDF>=1.21.1       # PDF processing
python-pptx>=0.6.21   # PowerPoint processing
openai>=1.3.0         # OpenAI API
requests>=2.28.1      # HTTP requests for APIs
pillow>=9.3.0         # Image processing
numpy>=1.23.5         # Numerical operations

# Audio processing
pydub>=0.25.1         # Audio file manipulation

# Video processing
moviepy>=1.0.3        # Video creation and editing

# Animation
matplotlib>=3.6.2     # For some animations and visualizations

# Utilities
tqdm>=4.64.1          # Progress bars
colorlog>=6.7.0       # Colored logging